feat: New UX for Creating Meme
- Adds new launcher activity
- Creates new single page layout with the necessary views
- Adds animation to initial three views

The Meme creation use case is pretty simple and can easily be split into multiple screens. The idea of the new UX is to make it conversational style where ther user need to focus on only one thing at a time. On a high level the user required to do,

1. Pick an image
2. Add a text at the top
3. Add a text at the bottom
4. Check preview of the meme
5. Save or Share it

The original UX had a single screen with these actions and attempting to direct the user by disabling few actions at the beginning and enabling it as the user performs initial actions. This could be improved by requiring the user to do one thing at a time and walk through him along the way.

Technically, We can create 2-3 activities or have a single activity with fragments/views to implement the UX. We can start with simple activity and add new activities as needed.